requirements.txt
# Core dependencies
fastapi>=0.95.0
uvicorn>=0.22.0
pydantic>=2.0.0
requests>=2.28.0
python-dotenv>=1.0.0
grpcio>=1.44.0
grpcio-tools>=1.44.0
protobuf>=3.19.0

# AI service client libraries
openai>=1.0.0
anthropic>=0.5.0
google-generativeai>=0.3.0

# Research APIs
PyGithub>=1.59.0
arxiv>=1.4.7
duckduckgo-search>=3.9.0

# For testing
pytest>=7.3.1
pytest-cov>=4.1.0

# For development
black>=23.3.0
isort>=5.12.0
mypy>=1.3.0
ruff>=0.0.262